things got messy
This might be a really good test of the version control capabilities of
git.  But basically I switched the empirical function so servo angle is
a function of the distance (this is inverse of what we were doing).  So
we get the servo angle based on the given target distance.  We might
have to cover our tracks more, because theta launch is quite hard to
calculate and we only use like 2 functions now… yeah.